Setup the dashboard bars (the bigsidebar and the smallsidebar) and the dashboard layout
The idea is on bigscreen the bigsidebar to be visible, but the smallsidebar - not, and on small screen - the bigsidebar not to be visible and the smallsidebar to be visible.
ONLY the initial structure of the big and small sidebar CSS files is setup and also only the initial structure of their jsx files.