Setup separate controller and router folders and files in order to avoid crowding to too much code in the server.js file
The logic (or the controllers) in server.js has been removed into jobController.js.

Notice! We replaced tere the controllers, which in this case are the callback
functions, but not the routes. Then we set separate file for the routes and delete all the methods with
routes and callback functions from here.

Now only the line app.use('/api/v1/jobs', jobRouter); replace them in server.js.