Setup validation middleware
Extract code from the server post request and set them in newly created middleware for the setup validation.

The withValidationErrors will be the same in the whole project.
We'll change validateTest (we'll pass different stuffs for check)

Or explained in another way - we set up this middleware and essentially for every request, we're just going to create a new instance and
in validateTest we'll just set up our conditions. We won't touch withValidationErrors.