Doing so that we could access the cookie
For that goal, we set up a package which is going to allow us to access the cookie and also we right away setup our first check in the authenticate user. We do this by installing the cookie parser package with npm i cookie-parser@1.4.6.

Now if i have the cookie - i can successfully access the jobs. If i am going to remove that cookie, it's going to be different. Then we won't have access to the specific route.