Verify JWT
As a result, we now have restricted access to the job routes. So we restrict the access if the cookie is not present or the JWT is not valid. However, if everything is fine, then we actually attach the user from the JWT to our request, which will allow us to set up the upcoming functionality. And this user object is going to be available in all of the controllers which have this authenticate user middleware sitting in front. In our case, those are all the job controllers.