Setup validation of the owner of the eventual request
Until now we didn't check whether the user who is making the request is the actual owner of the job. I.e. each user could access someone else's job. Now we added checks whether the concrete user is admin or owner of the request which fixed this.

Rename the variable for checking the id.