Start to setup the search form
By default, if there is no method post on the form in the HTML, actually browser makes the get request back to the same URL and provides input values as query string parameters.
It's not finished, we don't make request back to the database using the values from the URL (it will be done).